trying to get copilot set up
- inline_completion is a 0.12 feature from what I can tell so not sure
how to get it working with sidekick, I think I would need to figure out
how copilot.lua works.
That would leave me with two language servers, one that copilot.lua uses
and one that sidekick uses. Maybe not the end of the world

The NES suggestions are slow as shit

Sidekick doesn't start automatically I have to hit ctrl-. first for it
to work